Initial commit. Project already started.
Model.cpp/.h contain the Model class, which retains information about the frequencies of characters and provides a lookup table for characters based on frequency. It is mostly complete.

ArDecoder.cpp/.h contain the ArDecoder class, which provides an interface through which to decode messages. It is not complete.

ArEncoder.cpp/.h contain the ArEncoder class, which provides an interface through which to encode messages. It is not complete.