Partially working prototype.
Filled in put() and get(), which do the actual encoding.

The encoding and decoding seems to be currently working for frequency tables where every frequency is a power of 2. Does not seem to be working if there are frequencies that are not a power of 2.

For the moment, top has been changed to be inclusive, but other options are still being considered.

Code is very messy at the moment, and will be cleaned up after a fully-working prototype is achieved.